Split out convertFromContentBlockToRaw into its own file
For use in outside code
